pyproject.toml
[tool.poetry]
name = "bitrix-call-handler"
version = "0.1.0"
description = ""
authors = ["Kirill8769 <tkkrozn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
ffmpeg = "^1.4"
openai-whisper = "^20231117"
asyncpg = "^0.29.0"
types-requests = "^2.31.0.20240106"


[tool.poetry.group.lint.dependencies]
flake8 = "^6.1.0"
isort = "^5.12.0"
mypy = "^1.7.1"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.4"
pytest-cov = "^4.1.0"

[tool.isort]
line_length = 119

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = [".venv", "tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
